Allow passing standalone tokens and whitespaces to tokenizer
In order to parse generic grammars, the tokenizer must be flexible(
(i.e. configurable). This means it should allow the user to define
which tokens are standalone and which characters are whitespaces.

"Standalone tokens" refers to those tokens that will be split from the
containing tokens, for example "3+2" would be treated as 1 token if we
only take into account whitespaces. Depending on the specific grammar,
the character '-' can be standalone or not (it is standalone for C/C++,
java etc, but it is not for Lisp).